task5 -2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,6 +2,6 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-    float divided = toBeDivided/5;
+    double divided = (double)(toBeDivided)/5;
     System.out.println(divided);}
 }
